backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
alembic==1.18.4
Mako==1.3.10
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
PyMySQL==1.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7
mysql-connector-python==8.2.0
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
google-generativeai==0.3.0
openai==1.3.0
httpx==0.27.0
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
gunicorn==21.2.0
werkzeug==3.0.1
marshmallow==3.20.1
email-validator==2.1.0.post1
pytest==8.3.5
pytest-cov==6.1.1
